feat(app theme): redo app theming to a more scalable way.

+++ app/src/main/java/teka/android/organiks_platform_android/data/remote/retrofit/models/FruitCollection.kt
@@ -27,7 +27,7 @@ fun FruitCollectionRequest.toFruitCollection(): FruitCollectionEntity {
         uuid = this.collection_uuid,
         qty = this.quantity,
         date = this.collection_date,
-        fruitTypeId = this.fruit_type_id.toInt(),
+        fruitTypeId = this.fruit_type_id,
         createdAt = this.date
 
     )
